refactor 📦 use existing functions and check allocations
